move entire downgrade process into separate utils file
