+++ corehq/apps/accounting/tasks.py
@@ -23,8 +23,7 @@ from six.moves.urllib.parse import urlencode
 
 from corehq.apps.accounting.automated_reports import CreditsAutomatedReport
 from corehq.apps.accounting.utils.downgrade import (
-    is_subscription_eligible_for_downgrade_process,
-    send_downgrade_notice,
+    downgrade_eligible_domains,
 )
 from corehq.apps.accounting.utils.subscription import (
     assign_explicit_unpaid_subscription,
@@ -44,12 +43,9 @@ from corehq.apps.accounting.models import (
     BillingAccount,
     CreditLine,
     Currency,
-    DefaultProductPlan,
     DomainUserHistory,
-    EntryPoint,
     FeatureType,
     InvoicingPlan,
-    SoftwarePlanEdition,
     StripePaymentMethod,
     Subscription,
     SubscriptionAdjustment,
@@ -68,10 +64,6 @@ from corehq.apps.accounting.utils import (
     get_dimagi_from_email,
     log_accounting_error,
     log_accounting_info,
-)
-from corehq.apps.accounting.utils.invoicing import (
-    get_domains_with_subscription_invoices_over_threshold,
-    get_accounts_with_customer_invoices_over_threshold,
 )
 from corehq.apps.app_manager.dbaccessors import get_all_apps
 from corehq.apps.domain.models import Domain
@@ -685,115 +677,7 @@ def send_credits_on_hq_report():
 
 @periodic_task(run_every=crontab(minute=0, hour=9), queue='background_queue', acks_late=True)
 def run_downgrade_process():
-    today = datetime.date.today()
-
-    for domain, oldest_unpaid_invoice, total in get_domains_with_subscription_invoices_over_threshold(today):
-        current_subscription = Subscription.get_active_subscription_by_domain(domain)
-        if is_subscription_eligible_for_downgrade_process(current_subscription):
-            _apply_downgrade_process(oldest_unpaid_invoice, total, today, current_subscription)
-
-    for oldest_unpaid_invoice, total in get_accounts_with_customer_invoices_over_threshold(today):
-        subscription_on_invoice = oldest_unpaid_invoice.subscriptions.first()
-        if is_subscription_eligible_for_downgrade_process(subscription_on_invoice):
-            _apply_downgrade_process(oldest_unpaid_invoice, total, today)
-
-
-def _apply_downgrade_process(oldest_unpaid_invoice, total, today, subscription=None):
-    from corehq.apps.domain.views.accounting import DomainBillingStatementsView, DomainSubscriptionView
-    from corehq.apps.accounting.views import EnterpriseBillingStatementsView
-
-    context = {
-        'total': format(total, '7.2f'),
-        'date_60': oldest_unpaid_invoice.date_due + datetime.timedelta(days=60),
-        'contact_email': settings.INVOICING_CONTACT_EMAIL
-    }
-    if oldest_unpaid_invoice.is_customer_invoice:
-        domain = oldest_unpaid_invoice.subscriptions.first().subscriber.domain
-        context.update({
-            'statements_url': absolute_reverse(
-                EnterpriseBillingStatementsView.urlname, args=[domain]),
-            'domain_or_account': oldest_unpaid_invoice.account.name
-        })
-    else:
-        domain = subscription.subscriber.domain
-        context.update({
-            'domain': domain,
-            'subscription_url': absolute_reverse(DomainSubscriptionView.urlname,
-                                                 args=[domain]),
-            'statements_url': absolute_reverse(DomainBillingStatementsView.urlname,
-                                               args=[domain]),
-            'domain_or_account': domain
-        })
-
-    days_ago = (today - oldest_unpaid_invoice.date_due).days
-    if days_ago >= 61:
-        if not oldest_unpaid_invoice.is_customer_invoice:  # We do not automatically downgrade customer invoices
-            _downgrade_domain(subscription)
-            send_downgrade_notice(oldest_unpaid_invoice, context)
-    elif days_ago == 58:
-        _send_downgrade_warning(oldest_unpaid_invoice, context)
-    elif days_ago == 30:
-        _send_overdue_notice(oldest_unpaid_invoice, context)
-
-
-
-def _downgrade_domain(subscription):
-    subscription.change_plan(
-        DefaultProductPlan.get_default_plan_version(
-            SoftwarePlanEdition.PAUSED
-        ),
-        adjustment_method=SubscriptionAdjustmentMethod.AUTOMATIC_DOWNGRADE,
-        note='Automatic pausing of subscription for invoice 60 days late',
-        internal_change=True
-    )
-
-
-def _send_downgrade_warning(invoice, context):
-    if invoice.is_customer_invoice:
-        subject = _(
-            "CommCare Alert: {}'s subscriptions will be paused after tomorrow!".format(
-                invoice.account.name
-            ))
-        subscriptions_to_downgrade = _(
-            "subscriptions on {}".format(invoice.account.name)
-        )
-        bcc = None
-    else:
-        subject = _(
-            "CommCare Alert: {}'s subscription will be paused after tomorrow!".format(
-                invoice.get_domain()
-            ))
-        subscriptions_to_downgrade = _(
-            "subscription for {}".format(invoice.get_domain())
-        )
-        bcc = [settings.GROWTH_EMAIL]
-
-    context.update({
-        'subscriptions_to_downgrade': subscriptions_to_downgrade
-    })
-    send_html_email_async.delay(
-        subject,
-        invoice.contact_emails,
-        render_to_string('accounting/email/downgrade_warning.html', context),
-        render_to_string('accounting/email/downgrade_warning.txt', context),
-        cc=[settings.ACCOUNTS_EMAIL],
-        bcc=bcc,
-        email_from=get_dimagi_from_email())
-
-
-def _send_overdue_notice(invoice, context):
-    if invoice.is_customer_invoice:
-        bcc = None
-    else:
-        bcc = [settings.GROWTH_EMAIL]
-    send_html_email_async.delay(
-        _('CommCare Billing Statement 30 days Overdue for {}'.format(context['domain_or_account'])),
-        invoice.contact_emails,
-        render_to_string('accounting/email/30_days.html', context),
-        render_to_string('accounting/email/30_days.txt', context),
-        cc=[settings.ACCOUNTS_EMAIL],
-        bcc=bcc,
-        email_from=get_dimagi_from_email())
+    downgrade_eligible_domains()
 
 
 @task(serializer='pickle', queue='background_queue', ignore_result=True, acks_late=True,
